Massive refactor of pretty much everything
The way that state was being handled between different windows didn't
make much sense, so most of it has been refactored.

Now, the state is held by the main process and passed to each window.
The windows usually just want to listen to a specific slice (the part of
the state that describes that window) with the new connect utility.